requirements.txt
flask==2.3.3
sqllineage==1.5.1
python-dotenv==1.0.1
google-generativeai==0.8.4
gunicorn==21.2.0
